make user details with states, current profile api

+++ Backend/src/modules/users/users.controller.ts
@@ -3,7 +3,8 @@ import { usersService } from "./users.service";
 import { sendResponse } from "@/utils/sendResponse";
 import httpStatus from 'http-status'
 import type { RequestHandler } from "express";
-import type { Role, UserStatus } from "@/prisma/generated/prisma/enums";
+import { Role, type UserStatus } from "@/prisma/generated/prisma/enums";
+import ApiError from "@/errors/ApiError";
 
 export const getUsers: RequestHandler = catchAsync(async (req, res,): Promise<void> => {
     const { limit, cursor, search, role, status } = req.query;
@@ -30,7 +31,20 @@ export const getUsers: RequestHandler = catchAsync(async (req, res,): Promise<vo
 export const updateUser: RequestHandler = catchAsync(async (req, res) => {
     const { id } = req.params;
     const payload = req.body;
+    const loggedInUser = req.user;
 
+    if (!loggedInUser) {
+        throw new ApiError(httpStatus.UNAUTHORIZED, "Unauthorized access");
+    }
+    if (loggedInUser.role !== Role.ADMIN && loggedInUser.id !== id) {
+        throw new ApiError(httpStatus.FORBIDDEN, "Unauthorized to update this user!");
+    }
+
+
+    if (loggedInUser.role !== Role.ADMIN) {
+        delete payload.role;
+        delete payload.status;
+    }
     const user = await usersService.updateUserById(id as string, payload)
 
     sendResponse(res, {
@@ -43,14 +57,51 @@ export const updateUser: RequestHandler = catchAsync(async (req, res) => {
 })
 export const deleteUser: RequestHandler = catchAsync(async (req, res) => {
     const { id } = req.params;
+    const loggedInUser = req.user;
+
+    if (!loggedInUser) {
+        throw new ApiError(httpStatus.UNAUTHORIZED, "Unauthorized access");
+    }
+    if (loggedInUser.role !== Role.ADMIN) {
+        throw new ApiError(httpStatus.FORBIDDEN, "Unauthorized to delete this user!");
+    }
 
     const user = await usersService.deleteUserById(id as string)
 
     sendResponse(res, {
         statusCode: httpStatus.OK,
         success: true,
-        message: "Users Update successfully",
+        message: "Users Delete successfully",
         data: user,
-
     })
 })
+
+export const currentUserProfile: RequestHandler = catchAsync(async (req, res) => {
+    const loggedInUser = req.user;
+    if (!loggedInUser) {
+        throw new ApiError(httpStatus.UNAUTHORIZED, "Unauthorized access");
+    }
+    const user = await usersService.getCurrentUserProfile(loggedInUser.id)
+    sendResponse(res, {
+        statusCode: httpStatus.OK,
+        success: true,
+        message: "Fetch Current Profile successfully",
+        data: user,
+    })
+})
+
+export const userDetails: RequestHandler = catchAsync(async (req, res) => {
+    const { id: userId } = req.params;
+
+    if (!userId || typeof userId !== 'string') {
+        throw new ApiError(httpStatus.BAD_REQUEST, "Invalid or missing User ID");
+    }
+    
+    const result = await usersService.getUserDetailsWithStats(userId);
+    sendResponse(res, {
+        statusCode: httpStatus.OK,
+        success: true,
+        message: "Fetch User Details with states successfully",
+        data: result,
+    })
+})
